ref: Move formatSecondsToClock & parseClockToSeconds into utils/duration/
