+++ static/app/utils/duration/formatDuration.tsx
@@ -1,5 +1,5 @@
+import {formatSecondsToClock} from 'sentry/utils/duration/formatSecondsToClock';
 import type {Duration, Unit} from 'sentry/utils/duration/types';
-import {formatSecondsToClock} from 'sentry/utils/formatters';
 
 type Format =
   // example: `3,600`
