+++ static/app/utils/formatters.tsx
@@ -37,53 +37,6 @@ export const NANOSECOND = 0.000001;
  */
 export {getExactDuration} from 'sentry/utils/duration/getExactDuration';
 
-export function formatSecondsToClock(
-  seconds: number,
-  {padAll}: {padAll: boolean} = {padAll: true}
-) {
-  if (seconds === 0 || isNaN(seconds)) {
-    return padAll ? '00:00' : '0:00';
-  }
-
-  const divideBy = (msValue: number, time: number) => {
-    return {
-      quotient: msValue < 0 ? Math.ceil(msValue / time) : Math.floor(msValue / time),
-      remainder: msValue % time,
-    };
-  };
-
-  // value in milliseconds
-  const absMSValue = round(Math.abs(seconds * 1000));
-
-  const {quotient: hours, remainder: rMins} = divideBy(absMSValue, HOUR);
-  const {quotient: minutes, remainder: rSeconds} = divideBy(rMins, MINUTE);
-  const {quotient: secs, remainder: milliseconds} = divideBy(rSeconds, SECOND);
-
-  const fill = (num: number) => (num < 10 ? `0${num}` : String(num));
-
-  const parts = hours
-    ? [padAll ? fill(hours) : hours, fill(minutes), fill(secs)]
-    : [padAll ? fill(minutes) : minutes, fill(secs)];
-
-  const ms = `000${milliseconds}`.slice(-3);
-  return milliseconds ? `${parts.join(':')}.${ms}` : parts.join(':');
-}
-
-export function parseClockToSeconds(clock: string) {
-  const [rest, milliseconds] = clock.split('.');
-  const parts = rest.split(':');
-
-  let seconds = 0;
-  const progression = [MONTH, WEEK, DAY, HOUR, MINUTE, SECOND].slice(parts.length * -1);
-  for (let i = 0; i < parts.length; i++) {
-    const num = Number(parts[i]) || 0;
-    const time = progression[i] / 1000;
-    seconds += num * time;
-  }
-  const ms = Number(milliseconds) || 0;
-  return seconds + ms / 1000;
-}
-
 /**
  * Format a value between 0 and 1 as a percentage
  */
